Co-IntelligenceGenAIUniverse/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
boto3==1.34.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
streamlit==1.28.0
PyPDF2==3.0.1
python-docx==1.1.0
duckduckgo-search==6.1.12
